+++ packages/core/utils/device.d.ts
@@ -25,6 +25,8 @@ export interface Device {
   windows: boolean
   /** true when app running in cordova environment */
   cordova: boolean
+  /** true when app is running under Electron environment */
+  electron: boolean
   /** true when app running in cordova environment */
   phonegap: boolean
   /** true if app runs in web view - webapp installed to home screen */