+++ packages/react/components/link.js
@@ -85,7 +85,11 @@ class F7Link extends React.Component {
       id,
       style
     } = props;
-    const defaultSlots = self.slots.default;
+    const defaultSlots = self.slots.default || [];
+    Object.keys(self.slots).forEach(key => {
+      if (typeof self.slots[key] === 'undefined' || key === 'default') return;
+      self.slots[key].forEach(child => defaultSlots.push(child));
+    });
     let iconEl;
     let textEl;
     let badgeEl;