+++ packages/vue/components/popup.js
@@ -60,7 +60,7 @@ export default {
   },
 
   created() {
-    Utils.bindMethods(this, ['onOpen', 'onOpened', 'onClose', 'onClosed']);
+    Utils.bindMethods(this, ['onOpen', 'onOpened', 'onClose', 'onClosed', 'onSwipeStart', 'onSwipeMove', 'onSwipeEnd', 'onSwipeClose']);
   },
 
   mounted() {
@@ -80,6 +80,10 @@ export default {
     const popupParams = {
       el,
       on: {
+        swipeStart: self.onSwipeStart,
+        swipeMove: self.onSwipeMove,
+        swipeEnd: self.onSwipeEnd,
+        swipeClose: self.onSwipeClose,
         open: self.onOpen,
         opened: self.onOpened,
         close: self.onClose,
@@ -111,6 +115,22 @@ export default {
   },
 
   methods: {
+    onSwipeStart(instance) {
+      this.dispatchEvent('popup:swipestart popupSwipeStart', instance);
+    },
+
+    onSwipeMove(instance) {
+      this.dispatchEvent('popup:swipemove popupSwipeMove', instance);
+    },
+
+    onSwipeEnd(instance) {
+      this.dispatchEvent('popup:swipeend popupSwipeEnd', instance);
+    },
+
+    onSwipeClose(instance) {
+      this.dispatchEvent('popup:swipeclose popupSwipeClose', instance);
+    },
+
     onOpen(instance) {
       this.dispatchEvent('popup:open popupOpen', instance);
     },